Changed function prototypes for ext2 read/write to return 'ssize_t

+++ include/fs/ext2.h
@@ -240,10 +240,10 @@ ext2_block ext2_bmap(struct ext2_inode_m *ip, off_t off);
 ext2_block ext2_create_block(struct ext2_inode_m *ip, off_t off);
 
 void *ext2_open(const char *path, int flags, int mode, int *err);
-int ext2_read(struct file *file, void *buf, size_t count);
-int ext2_readi(void *inode, void *buf, size_t count, off_t *off);
-int ext2_write(struct file *file, void *buf, size_t count);
-int ext2_writei(void *inode, void *buf, size_t count, off_t *off);
+ssize_t ext2_read(struct file *file, void *buf, size_t count);
+ssize_t ext2_readi(void *inode, void *buf, size_t count, off_t *off);
+ssize_t ext2_write(struct file *file, void *buf, size_t count);
+ssize_t ext2_writei(void *inode, void *buf, size_t count, off_t *off);
 int ext2_close(struct file *file);
 
 int ext2_permission(struct ext2_inode_m *ip, enum ext2_perm_mask mask);
